pyproject.toml
[tool.poetry]
name = "gigs"
version = "0.1.0"
description = ""
authors = ["mintydevdaz <mintydevdaz@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.24.1"
selectolax = "^0.3.16"
pydantic = "^2.1.1"
chompjs = "^1.2.2"
python-dotenv = "^1.0.0"
polars = "^0.19.2"
pretty-html-table = "^0.9.16"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
